+++ run/fabric.sh
@@ -4,11 +4,12 @@ WORKING_DIR=${PWD##*/}
 WORKING_DIR=${WORKING_DIR:-/}
 if [ $WORKING_DIR == "run" ]; then
     cd ../;
+    echo "Going up a directory";
 fi
 ./run/fabric/stop.sh
 
-gradle :fabric:build
-cp ./fabric/build/libs/fabric.jar ./run/fabric/mods/fabric-discord.jar
+gradle platforms:fabric:build
+cp ./platforms/fabric/build/libs/fabric-all.jar ./run/fabric/mods/fabric-discord.jar
 
 cd ./run/fabric
 # Disable the proxy mod