+++ run/velocity.sh
@@ -5,13 +5,13 @@ WORKING_DIR=${WORKING_DIR:-/}
 if [ $WORKING_DIR == "run" ]; then
     cd ../;
 fi
-gradle :velocity:build
+gradle platforms:velocity:build
 
 cd ./run
 ./velocity/stop.sh
 ./fabric/stop.sh
 
-cp ../velocity/build/libs/velocity-all.jar ./velocity/plugins/velocity-discord.jar
+cp ../platforms/velocity/build/libs/velocity-all.jar ./velocity/plugins/velocity-discord.jar
 
 # Disable the discord mod on the fabric server, enable the proxy mod
 mv ./fabric/mods/FabricProxy-Lite.jar.disabled ./fabric/mods/FabricProxy-Lite.jar